LPS-79195 Revert the ordering implemented by LPS-50481 and instead do not delete groups that are the staging group. With the ordering restored, deleting instances where an Organization Site has a child site completes successfully.

+++ portal-impl/src/com/liferay/portal/service/impl/CompanyLocalServiceImpl.java
@@ -1345,16 +1345,6 @@ public class CompanyLocalServiceImpl extends CompanyLocalServiceBaseImpl {
 
 		accountLocalService.deleteAccount(company.getAccountId());
 
-		// Organizations
-
-		DeleteOrganizationActionableDynamicQuery
-			deleteOrganizationActionableDynamicQuery =
-				new DeleteOrganizationActionableDynamicQuery();
-
-		deleteOrganizationActionableDynamicQuery.setCompanyId(companyId);
-
-		deleteOrganizationActionableDynamicQuery.performActions();
-
 		// Groups
 
 		DeleteGroupActionableDynamicQuery deleteGroupActionableDynamicQuery =
@@ -1418,6 +1408,16 @@ public class CompanyLocalServiceImpl extends CompanyLocalServiceBaseImpl {
 			});
 
 		layoutSetPrototypeActionableDynamicQuery.performActions();
+
+		// Organizations
+
+		DeleteOrganizationActionableDynamicQuery
+			deleteOrganizationActionableDynamicQuery =
+				new DeleteOrganizationActionableDynamicQuery();
+
+		deleteOrganizationActionableDynamicQuery.setCompanyId(companyId);
+
+		deleteOrganizationActionableDynamicQuery.performActions();
 
 		// Roles
 
@@ -1737,7 +1737,7 @@ public class CompanyLocalServiceImpl extends CompanyLocalServiceBaseImpl {
 						throws PortalException {
 
 						if (!PortalUtil.isSystemGroup(group.getGroupKey()) &&
-							!group.isCompany()) {
+							!group.isCompany() && !group.isStagingGroup()) {
 
 							deleteGroup(group);
 						}
